pyproject.toml
[project]
name = "libroapi"
version = "0.1.0"
description = ""
authors = [
    {name = "dineshkumarkarimajji-bootlabs",email = "dineshkumar.karimajji@bootlabstech.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.12.0,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib (>=0.2.1,<0.3.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.2)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
